refactor: EventHandler 인터페이스를 별도의 파일로 분리
