+++ app/src/main/java/woowacourse/shopping/view/detail/ProductDetailEventHandler.kt
@@ -1,0 +1,7 @@
+package woowacourse.shopping.view.detail
+
+import woowacourse.shopping.view.model.InventoryItem
+
+interface ProductDetailEventHandler {
+    fun onAddToCartSelected(product: InventoryItem.ProductUiModel)
+}
