+++ app/src/main/java/woowacourse/shopping/view/inventory/InventoryEventHandler.kt
@@ -1,0 +1,9 @@
+package woowacourse.shopping.view.inventory
+
+import woowacourse.shopping.view.model.InventoryItem
+
+interface InventoryEventHandler {
+    fun onProductSelected(product: InventoryItem.ProductUiModel)
+
+    fun onLoadMoreProducts()
+}
