+++ app/src/main/java/woowacourse/shopping/view/shoppingcart/ShoppingCartEventHandler.kt
@@ -1,0 +1,11 @@
+package woowacourse.shopping.view.shoppingcart
+
+import woowacourse.shopping.domain.Product
+
+interface ShoppingCartEventHandler {
+    fun onProductRemove(product: Product)
+
+    fun onPaginationPrevious()
+
+    fun onPaginationNext()
+}
